# Conflicts:
#	.idea/compiler.xml
#	src/main/java/com/jxd/studentmanager/controller/StudentController.java

+++ src/main/java/com/jxd/studentmanager/controller/EmpController.java
@@ -117,6 +117,7 @@ public class EmpController {
         }
     }
 
+
     /**
      * 修改一个员工的信息
      * @param emp
@@ -162,38 +163,54 @@ public class EmpController {
     public String delEmpByIdCascade(int eid) {
         boolean flag = false;
 
-        QueryWrapper<User> wrapper = new QueryWrapper<>();
-        wrapper.eq("eid",eid);
-        User user = userService.getOne(wrapper);
-
-        if(user.getRole() == 0){
-            //管理员账号
-            flag = userService.removeById(user.getUid());
-            flag = empService.removeById(eid);
-
-        } else if (user.getRole() == 3){
-            //学生账号
-            QueryWrapper<Student> wrapper1 = new QueryWrapper<>();
-            wrapper1.eq("eid",eid);
-            Student student = studentService.getOne(wrapper1);
-
-            UpdateWrapper<StudentScore> wrapper2 = new UpdateWrapper<>();
-            wrapper2.eq("sid",student.getSid());
-            flag = studentScoreService.remove(wrapper2);
-
-            flag = studentService.removeById(student.getEid());
-        } else {
-            //教师或者部门经理账号
+        Emp emp1 = empService.getById(eid);
+        if(emp1.getIsdel() == 1){
             QueryWrapper<StudentScore> wrapper1 = new QueryWrapper<>();
             wrapper1.eq("eid",eid);
             int num = studentScoreService.count(wrapper1);
-
             if(num == 0){
                 flag = empService.removeById(eid);
+            }
+            return "success";
+        }
+
+        QueryWrapper<User> wrapper = new QueryWrapper<>();
+        wrapper.eq("uname",eid);
+        User user = userService.getOne(wrapper);
+
+        if(user != null){
+            if(user.getRole() == 0){
+                //管理员账号
+                flag = userService.removeById(user.getUid());
+                flag = empService.removeById(eid);
+
+            } else if (user.getRole() == 3){
+                //学生账号
+                QueryWrapper<Student> wrapper1 = new QueryWrapper<>();
+                wrapper1.eq("eid",eid);
+                Student student = studentService.getOne(wrapper1);
+
+                UpdateWrapper<StudentScore> wrapper2 = new UpdateWrapper<>();
+                wrapper2.eq("sid",student.getSid());
+                flag = studentScoreService.remove(wrapper2);
+
+                flag = studentService.removeById(student.getSid());
+                flag = userService.removeById(user.getUid());
+                flag = empService.removeById(eid);
             } else {
-                Emp emp = empService.getById(eid);
-                emp.setIsdel(1);
-                flag = empService.updateById(emp);
+                //教师或者部门经理账号
+                QueryWrapper<StudentScore> wrapper1 = new QueryWrapper<>();
+                wrapper1.eq("eid",eid);
+                int num = studentScoreService.count(wrapper1);
+
+                if(num == 0){
+                    flag = empService.removeById(eid);
+                } else {
+                    Emp emp = empService.getById(eid);
+                    emp.setIsdel(1);
+                    flag = empService.updateById(emp);
+                }
+                userService.removeById(user.getUid());
             }
         }
         if(flag){
@@ -223,7 +240,7 @@ public class EmpController {
     @ResponseBody
     public String delEmpsByIdsCascade(int[] eids) {
         for(int eid : eids){
-            String result = delEmpById(eid);
+            String result = delEmpByIdCascade(eid);
             if(result.equals("fail")){
                 return "fail";
             }
@@ -240,6 +257,7 @@ public class EmpController {
         }
         return "success";
     }
+
 
     @RequestMapping("/getManager")
     @ResponseBody
